+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
  - PHP 8 and 8.1 support
  - Latest postmen.com JSON schema changes
+ - Support async requests (label creation for webhook support)
 
 ### Removed
  - PHP 7.1 support